Update EZMultiPliDeviceType.gy
Changed to the new ST FingerPrint method which is required to be unique for all certified Z-Wave devices vs. the old method of comparing the command classes which was definitely not a good algorithm.
Lowered the parameter settings to be more "in line with ST user expectations" according to Don Kirker of ST.
Can't go too low or you'll swamp the Z-Wave network with too much useless traffic.